Convert Proficiency.base from Integer to Float.
NOTE: may have unintended consequences.

+++ abstraction_lv1/proficiencies_template.py
@@ -38,7 +38,7 @@ class Proficiency(TemplateMixin, Base):
 
     # Main colums
     level = Column(Integer)
-    base = Column(Integer)
+    base = Column(Float)
     modifier = Column(Float)
 
     type_ = Column(String(50))
